define pytest_plugins in the top-level conftest
